move extensions to persistence specific files
